+++ src-tauri/src/parsers/dmm/mod.rs
@@ -1,0 +1,10 @@
+//! DMM通販パーサーモジュール
+//!
+//! 各パーサーは dmm/ サブモジュールとして配置されています。
+
+pub mod cancel;
+pub mod confirm;
+pub mod merge_complete;
+pub mod order_number_change;
+pub mod send;
+pub mod split_complete;
