+++ src-tauri/src/parsers/hobbysearch/cancel.rs
