+++ src-tauri/src/parsers/hobbysearch/change.rs
@@ -1,5 +1,5 @@
-use super::hobbysearch_common::{extract_amounts, extract_delivery_address, parse_item_line};
-use super::{EmailParser, OrderInfo, OrderItem};
+use super::{extract_amounts, extract_delivery_address, parse_item_line};
+use crate::parsers::{EmailParser, OrderInfo, OrderItem};
 use regex::Regex;
 
 /// 組み換え（購入分）メール用パーサー
