+++ src-tauri/src/parsers/hobbysearch/send.rs
@@ -1,7 +1,5 @@
-use super::hobbysearch_common::{
-    extract_amounts, extract_delivery_address, extract_delivery_info, parse_item_line,
-};
-use super::{EmailParser, OrderInfo, OrderItem};
+use super::{extract_amounts, extract_delivery_address, extract_delivery_info, parse_item_line};
+use crate::parsers::{EmailParser, OrderInfo, OrderItem};
 use regex::Regex;
 
 /// 発送通知メール用パーサー
